+++ src/main/scala/org/kuleuven/esat/graphUtils/Point.scala
@@ -9,16 +9,16 @@ import com.tinkerpop.frames.{Adjacency, VertexFrame, Property}
  */
 trait Point extends VertexFrame {
   @Property("value")
-  def getValue(): Array[Byte]
+  def getValue(): Array[Double]
 
   @Property("value")
-  def setValue(value: Array[Byte]): Unit
+  def setValue(value: Array[Double]): Unit
 
   @Property("featureMap")
-  def getFeatureMap(): Array[Byte]
+  def getFeatureMap(): Array[Double]
 
   @Property("featureMap")
-  def setFeatureMap(value: Array[Byte]): Unit
+  def setFeatureMap(value: Array[Double]): Unit
 
   @Adjacency(label = "causes", direction = Direction.OUT)
   def getLabel(): java.lang.Iterable[Label]
